// types/driver.ts
export interface Driver {
  id: string;
  fullName: string;
  givenName: string;
  familyName: string;
  teamId: string;
  teamName: string;
  teamPoints: number;
  teamColor: string;
  number: number;
  position: number;
  points: number;
  wins: number;
  podiums: number;
  nationality: string;
  dateOfBirth: string;
  profile: string; 
}